Node Reverse Engineering
I want to help contribute to this project, so I'm going to reverse engineer how it works, and possibly migrate it to TypeScript so it can be used from other modules safely as well. I may also shamelessly plug/use my NBTify library in the process, I thought that could work really well for this project in ESM land in the browser. Things could even be set up with Vite here, but that may not be necessary.

I'm going to try building things out in TS to start, since that's easiest to migrate with at times, then I will come back and annotate it with JSDoc types instead, if that's more akin to the likes of what this project is focused around. But, we'll see. TS is wonderous, the compile step really allows you to use it anywhere! No need to handle multiple module systems in your single codebase, the compile step will manage that hassle for you, since you choose which targets you'd like to compile down to.