Dependency Tracking
Going to start managing the dependencies using npm rather than with individual window scripts. This will also make it easier to migrate to NBTify eventually, in addition to using the Compression Streams API with NBTify in favor of a `node:zlib` polyfill/browser fallback. Nowadays you can use a single, cross-platform API which is available on all platforms, be it Node or the browsers :)