Module-Scoped Function Implementations
Now that things are scoped to the ESM module, I can move the implementation functions out of the `schemtoschematic` definition, and out in the open to be declared only a single time, rather than being redefined with each call of `schemtoschematic`.

The diff looks like I moved the main function body, but that's because I moved all of the other content out, and didn't indent it to match yet. I only like to show diffs as local to the changes as they can be. If you can't see what changed in the diff, then what's the point of it? :P