TileEntity Mapping Logic
This appeared to be set up a little differently than intended, I think this is because of how NBT.js handles ListTag values, it's a bit inconsistent in how one is supposed to access the values and properties out of it.

I tried implementing NBTify before this fix, and it's how I encountered finding this. I wanted to go ahead and fix this handling before moving the library backing over, since this is specific to the implementation of the converter, and now related to the library. I want to ensure moving to NBTify produces 1:1 outputs with the original converter, and this ensures that the original output is what it's meant to be, and will also align with NBTify.