Mutable to Immutable Pt.1
Not completely sure about the NBT structure for the older `.schematic` format, so I'm going to try digging into that a little more before moving things over completely. This is the first test at moving things over, but you can see that the file size in this commit is different than that of the original file, so some of the needed keys appear to be missing. Unless the original version of this script included extra ones that aren't necessary anymore. Either way, I want things to be type-checked properly.

https://www.reddit.com/r/Minecraft/comments/ohabcm/convert_schem_file_to_schematic/
https://old.reddit.com/r/Minecraft/comments/oabwea/schem_files_work_in_litematica_but_schematic/h3ju3sy/
https://github.com/SpongePowered/Schematic-Specification